\name{RCLS.chunk-methods}
\docType{methods}
\alias{g.levels}
\alias{g.train}
\alias{g.test}
\alias{g.s,RCLS.chunk-method}
\alias{g.levels,RCLS.chunk-method}
\alias{g.ntrain,RCLS.chunk-method}
\alias{g.train,RCLS.chunk-method}
\alias{g.ntest,RCLS.chunk-method}
\alias{g.test,RCLS.chunk-method}
\alias{g.Zt,RCLS.chunk-method}
\title{
?
}
\description{
?
}
\author{Marko Nagode}
\keyword{classification}

