\name{rebmix-internal}
\alias{AIC,ANY-method}
\alias{AIC3,ANY-method}
\alias{AIC4,ANY-method}
\alias{AICc,ANY-method}
\alias{CAIC,ANY-method}
\alias{AWE,ANY-method}
\alias{BIC,ANY-method}
\alias{CLC,ANY-method}
\alias{HQC,ANY-method}
\alias{ICL,ANY-method}
\alias{ICLBIC,ANY-method}
\alias{logL,ANY-method}
\alias{MDL2,ANY-method}
\alias{MDL5,ANY-method}
\alias{PC,ANY-method}
\alias{PRD,ANY-method}
\alias{SSE,ANY-method}
\alias{REBMIX,ANY-method}
\alias{RNGMIX,ANY-method}
\alias{RCLRMIX,ANY-method}
\alias{RCLSMIX,ANY-method}
\alias{BFSMIX,ANY-method}
\alias{EMMIX,ANY-method}
\alias{fhistogram,ANY-method}
\alias{chistogram,ANY-method}
\alias{mapclusters,ANY-method}
\alias{labelmoments,ANY-method}
\alias{mergelabels,ANY-method}

\alias{a.d}
\alias{a.theta1<-}
\alias{a.theta2<-}
\alias{a.theta3<-}
\alias{a.c,RNGMIX.Theta-method}
\alias{a.d,RNGMIX.Theta-method}
\alias{a.pdf,RNGMIX.Theta-method}
\alias{a.Theta,RNGMIX.Theta-method}
\alias{a.theta1<-,RNGMIX.Theta,ANY-method}
\alias{a.theta2<-,RNGMIX.Theta,ANY-method}
\alias{a.theta3<-,RNGMIX.Theta,ANY-method}
\alias{a.theta1<-,RNGMIX.Theta,missing-method}
\alias{a.theta2<-,RNGMIX.Theta,missing-method}
\alias{a.theta3<-,RNGMIX.Theta,missing-method}
\alias{a.c,RNGMVNORM.Theta-method}
\alias{a.d,RNGMVNORM.Theta-method}
\alias{a.pdf,RNGMVNORM.Theta-method}
\alias{a.Theta,RNGMVNORM.Theta-method}
\alias{a.theta1<-,RNGMVNORM.Theta,ANY-method}
\alias{a.theta2<-,RNGMVNORM.Theta,ANY-method}
\alias{a.theta1.all<-,RNGMIX.Theta-method}
\alias{a.theta2.all<-,RNGMIX.Theta-method}
\alias{a.theta3.all<-,RNGMIX.Theta-method}
\alias{a.theta1.all<-,RNGMVNORM.Theta-method}
\alias{a.theta2.all<-,RNGMVNORM.Theta-method}

\alias{a.Dataset.name}
\alias{a.rseed}
\alias{a.n}
\alias{a.Theta}
\alias{a.Dataset}
\alias{a.Zt}
\alias{a.w}
\alias{a.Variables}
\alias{a.ymin}
\alias{a.ymax}
\alias{a.Dataset.name,RNGMIX-method}
\alias{a.rseed,RNGMIX-method}
\alias{a.n,RNGMIX-method}
\alias{a.Theta,RNGMIX-method}
\alias{a.Dataset,RNGMIX-method}
\alias{a.Zt,RNGMIX-method}
\alias{a.w,RNGMIX-method}
\alias{a.Variables,RNGMIX-method}
\alias{a.ymin,RNGMIX-method}
\alias{a.ymax,RNGMIX-method}
\alias{a.Dataset.name,RNGMVNORM-method}
\alias{a.rseed,RNGMVNORM-method}
\alias{a.n,RNGMVNORM-method}
\alias{a.Theta,RNGMVNORM-method}
\alias{a.Dataset,RNGMVNORM-method}
\alias{a.Zt,RNGMVNORM-method}
\alias{a.w,RNGMVNORM-method}
\alias{a.Variables,RNGMVNORM-method}
\alias{a.ymin,RNGMVNORM-method}
\alias{a.ymax,RNGMVNORM-method}

\alias{a.Preprocessing}
\alias{a.cmax}
\alias{a.cmin}
\alias{a.Criterion}
\alias{a.pdf}
\alias{a.theta1}
\alias{a.theta2}
\alias{a.theta3}
\alias{a.theta1.all}
\alias{a.theta2.all}
\alias{a.theta3.all}
\alias{a.theta1.all<-}
\alias{a.theta2.all<-}
\alias{a.theta3.all<-}
\alias{a.K}
\alias{a.y0}
\alias{a.ar}
\alias{a.Restraints}
\alias{a.Mode}
\alias{a.summary}
\alias{a.pos}
\alias{a.opt.c}
\alias{a.opt.IC}
\alias{a.opt.logL}
\alias{a.opt.Dmin}
\alias{a.opt.D}
\alias{a.all.K}
\alias{a.all.IC}
\alias{a.Dataset,REBMIX-method}
\alias{a.Preprocessing,REBMIX-method}
\alias{a.cmax,REBMIX-method}
\alias{a.cmin,REBMIX-method}
\alias{a.Criterion,REBMIX-method}
\alias{a.Variables,REBMIX-method}
\alias{a.pdf,REBMIX-method}
\alias{a.theta1,REBMIX-method}
\alias{a.theta2,REBMIX-method}
\alias{a.theta3,REBMIX-method}
\alias{a.theta1.all,REBMIX-method}
\alias{a.theta2.all,REBMIX-method}
\alias{a.theta3.all,REBMIX-method}
\alias{a.theta1.all,REBMVNORM-method}
\alias{a.theta2.all,REBMVNORM-method}
\alias{a.K,REBMIX-method}
\alias{a.ymin,REBMIX-method}
\alias{a.ymax,REBMIX-method}
\alias{a.ar,REBMIX-method}
\alias{a.Restraints,REBMIX-method}
\alias{a.Mode,REBMIX-method}
\alias{a.w,REBMIX-method}
\alias{a.Theta,REBMIX-method}
\alias{a.summary,REBMIX-method}
\alias{a.pos,REBMIX-method}
\alias{a.opt.c,REBMIX-method}
\alias{a.opt.IC,REBMIX-method}
\alias{a.opt.logL,REBMIX-method}
\alias{a.opt.Dmin,REBMIX-method}
\alias{a.opt.D,REBMIX-method}
\alias{a.all.K,REBMIX-method}
\alias{a.all.IC,REBMIX-method}
\alias{a.Dataset,REBMVNORM-method}
\alias{a.Preprocessing,REBMVNORM-method}
\alias{a.cmax,REBMVNORM-method}
\alias{a.cmin,REBMVNORM-method}
\alias{a.Criterion,REBMVNORM-method}
\alias{a.Variables,REBMVNORM-method}
\alias{a.pdf,REBMVNORM-method}
\alias{a.theta1,REBMVNORM-method}
\alias{a.theta2,REBMVNORM-method}
\alias{a.K,REBMVNORM-method}
\alias{a.ymin,REBMVNORM-method}
\alias{a.ymax,REBMVNORM-method}
\alias{a.ar,REBMVNORM-method}
\alias{a.Restraints,REBMVNORM-method}
\alias{a.Mode,REBMVNORM-method}
\alias{a.w,REBMVNORM-method}
\alias{a.Theta,REBMVNORM-method}
\alias{a.summary,REBMVNORM-method}
\alias{a.pos,REBMVNORM-method}
\alias{a.opt.c,REBMVNORM-method}
\alias{a.opt.IC,REBMVNORM-method}
\alias{a.opt.logL,REBMVNORM-method}
\alias{a.opt.Dmin,REBMVNORM-method}
\alias{a.opt.D,REBMVNORM-method}
\alias{a.all.K,REBMVNORM-method}
\alias{a.all.IC,REBMVNORM-method}

\alias{a.Bootstrap}
\alias{a.B}
\alias{a.replace}
\alias{a.prob}
\alias{a.c}
\alias{a.c.se}
\alias{a.c.cv}
\alias{a.c.mode}
\alias{a.c.prob}
\alias{a.w.se}
\alias{a.w.cv}
\alias{a.Theta.se}
\alias{a.Theta.cv}
\alias{a.rseed,REBMIX.boot-method}
\alias{a.pos,REBMIX.boot-method}
\alias{a.Bootstrap,REBMIX.boot-method}
\alias{a.B,REBMIX.boot-method}
\alias{a.n,REBMIX.boot-method}
\alias{a.replace,REBMIX.boot-method}
\alias{a.prob,REBMIX.boot-method}
\alias{a.c,REBMIX.boot-method}
\alias{a.c.se,REBMIX.boot-method}
\alias{a.c.cv,REBMIX.boot-method}
\alias{a.c.mode,REBMIX.boot-method}
\alias{a.c.prob,REBMIX.boot-method}
\alias{a.w,REBMIX.boot-method}
\alias{a.w.se,REBMIX.boot-method}
\alias{a.w.cv,REBMIX.boot-method}
\alias{a.Theta,REBMIX.boot-method}
\alias{a.Theta.se,REBMIX.boot-method}
\alias{a.Theta.cv,REBMIX.boot-method}
\alias{a.rseed,REBMVNORM.boot-method}
\alias{a.pos,REBMVNORM.boot-method}
\alias{a.Bootstrap,REBMVNORM.boot-method}
\alias{a.B,REBMVNORM.boot-method}
\alias{a.n,REBMVNORM.boot-method}
\alias{a.replace,REBMVNORM.boot-method}
\alias{a.prob,REBMVNORM.boot-method}
\alias{a.c,REBMVNORM.boot-method}
\alias{a.c.se,REBMVNORM.boot-method}
\alias{a.c.cv,REBMVNORM.boot-method}
\alias{a.c.mode,REBMVNORM.boot-method}
\alias{a.c.prob,REBMVNORM.boot-method}
\alias{a.w,REBMVNORM.boot-method}
\alias{a.w.se,REBMVNORM.boot-method}
\alias{a.w.cv,REBMVNORM.boot-method}
\alias{a.Theta,REBMVNORM.boot-method}
\alias{a.Theta.se,REBMVNORM.boot-method}
\alias{a.Theta.cv,REBMVNORM.boot-method}

\alias{a.Zp}
\alias{a.p}
\alias{a.pi}
\alias{a.P}
\alias{a.tau}
\alias{a.Rule}
\alias{a.from}
\alias{a.to}
\alias{a.EN}
\alias{a.ED}
\alias{a.Dataset,RCLRMIX-method}
\alias{a.pos,RCLRMIX-method}
\alias{a.Zt,RCLRMIX-method}
\alias{a.Zp,RCLRMIX-method}
\alias{a.c,RCLRMIX-method}
\alias{a.p,RCLRMIX-method}
\alias{a.pi,RCLRMIX-method}
\alias{a.P,RCLRMIX-method}
\alias{a.tau,RCLRMIX-method}
\alias{a.prob,RCLRMIX-method}
\alias{a.Rule,RCLRMIX-method}
\alias{a.from,RCLRMIX-method}
\alias{a.to,RCLRMIX-method}
\alias{a.EN,RCLRMIX-method}
\alias{a.ED,RCLRMIX-method}
\alias{a.Dataset,RCLRMVNORM-method}
\alias{a.pos,RCLRMVNORM-method}
\alias{a.Zt,RCLRMVNORM-method}
\alias{a.Zp,RCLRMVNORM-method}
\alias{a.c,RCLRMVNORM-method}
\alias{a.p,RCLRMVNORM-method}
\alias{a.pi,RCLRMVNORM-method}
\alias{a.P,RCLRMVNORM-method}
\alias{a.tau,RCLRMVNORM-method}
\alias{a.prob,RCLRMVNORM-method}
\alias{a.Rule,RCLRMVNORM-method}
\alias{a.from,RCLRMVNORM-method}
\alias{a.to,RCLRMVNORM-method}
\alias{a.EN,RCLRMVNORM-method}
\alias{a.ED,RCLRMVNORM-method}

\alias{a.levels}
\alias{a.train}
\alias{a.Zr}
\alias{a.test}
\alias{a.s,RCLS.chunk-method}
\alias{a.levels,RCLS.chunk-method}
\alias{a.ntrain,RCLS.chunk-method}
\alias{a.train,RCLS.chunk-method}
\alias{a.Zr,RCLS.chunk-method}
\alias{a.ntest,RCLS.chunk-method}
\alias{a.test,RCLS.chunk-method}
\alias{a.Zt,RCLS.chunk-method}

\alias{a.o}
\alias{a.s}
\alias{a.ntrain}
\alias{a.ntest}
\alias{a.CM}
\alias{a.Accuracy}
\alias{a.Error}
\alias{a.Precision}
\alias{a.Sensitivity}
\alias{a.Specificity}
\alias{a.Chunks}
\alias{a.o,RCLSMIX-method}
\alias{a.Dataset,RCLSMIX-method}
\alias{a.s,RCLSMIX-method}
\alias{a.ntrain,RCLSMIX-method}
\alias{a.P,RCLSMIX-method}
\alias{a.ntest,RCLSMIX-method}
\alias{a.Zt,RCLSMIX-method}
\alias{a.Zp,RCLSMIX-method}
\alias{a.CM,RCLSMIX-method}
\alias{a.Accuracy,RCLSMIX-method}
\alias{a.Error,RCLSMIX-method}
\alias{a.Precision,RCLSMIX-method}
\alias{a.Sensitivity,RCLSMIX-method}
\alias{a.Specificity,RCLSMIX-method}
\alias{a.Chunks,RCLSMIX-method}
\alias{a.o,RCLSMVNORM-method}
\alias{a.Dataset,RCLSMVNORM-method}
\alias{a.s,RCLSMVNORM-method}
\alias{a.ntrain,RCLSMVNORM-method}
\alias{a.P,RCLSMVNORM-method}
\alias{a.ntest,RCLSMVNORM-method}
\alias{a.Zt,RCLSMVNORM-method}
\alias{a.Zp,RCLSMVNORM-method}
\alias{a.CM,RCLSMVNORM-method}
\alias{a.Accuracy,RCLSMVNORM-method}
\alias{a.Error,RCLSMVNORM-method}
\alias{a.Precision,RCLSMVNORM-method}
\alias{a.Sensitivity,RCLSMVNORM-method}
\alias{a.Specificity,RCLSMVNORM-method}
\alias{a.Chunks,RCLSMVNORM-method}

\alias{a.strategy}
\alias{a.variant}
\alias{a.acceleration}
\alias{a.tolerance}
\alias{a.acceleration.multiplier}
\alias{a.maximum.iterations}
\alias{a.eliminate.zero.components}
\alias{a.likelihood.estimation.rule}
\alias{a.likelihood.tolerance.check}

\alias{a.strategy,EM.Control-method}
\alias{a.variant,EM.Control-method}
\alias{a.acceleration,EM.Control-method}
\alias{a.tolerance,EM.Control-method}
\alias{a.acceleration.multiplier,EM.Control-method}
\alias{a.maximum.iterations,EM.Control-method}
\alias{a.eliminate.zero.components,EM.Control-method}
\alias{a.likelihood.estimation.rule,EM.Control-method}
\alias{a.likelihood.tolerance.check,EM.Control-method}
\alias{a.K,EM.Control-method}

\alias{a.strategy<-}
\alias{a.variant<-}
\alias{a.acceleration<-}
\alias{a.tolerance<-}
\alias{a.acceleration.multiplier<-}
\alias{a.maximum.iterations<-}
\alias{a.eliminate.zero.components<-}
\alias{a.likelihood.estimation.rule<-}
\alias{a.likelihood.tolerance.check<-}
\alias{a.K<-}

\alias{a.strategy<-,EM.Control-method}
\alias{a.variant<-,EM.Control-method}
\alias{a.acceleration<-,EM.Control-method}
\alias{a.tolerance<-,EM.Control-method}
\alias{a.acceleration.multiplier<-,EM.Control-method}
\alias{a.maximum.iterations<-,EM.Control-method}
\alias{a.eliminate.zero.components<-,EM.Control-method}
\alias{a.likelihood.estimation.rule<-,EM.Control-method}
\alias{a.likelihood.tolerance.check<-,EM.Control-method}
\alias{a.K<-,EM.Control-method}

\alias{a.summary.EM}
\alias{a.summary.EM,REBMIX-method}
\alias{a.summary.EM,REBMVNORM-method}

\alias{a.w,EMMIX.Theta-method}
\alias{a.w,EMMVNORM.Theta-method}
\alias{a.w<-}
\alias{a.w<-,EMMIX.Theta-method}
\alias{a.w<-,EMMVNORM.Theta-method}

\alias{a.Y}
\alias{a.Y<-}
\alias{a.h}
\alias{a.ns}
\alias{a.Y,Histogram-method}
\alias{a.Y<-,Histogram-method}
\alias{a.K,Histogram-method}
\alias{a.ymin,Histogram-method}
\alias{a.ymax,Histogram-method}
\alias{a.y0,Histogram-method}
\alias{a.h,Histogram-method}
\alias{a.n,Histogram-method}
\alias{a.ns,Histogram-method}

\title{Internal rebmix Functions, Methods and Classes}
\description{
Internal rebmix functions, methods and classes.
}
\details{
These are not to be called by the user.
}
\keyword{internal}
