\name{THETA-methods}
\alias{THETA}
\alias{THETA-methods}
\alias{show,THETA-method}
\alias{g.d}
\alias{g.Theta1<-}
\alias{g.Theta2<-}
\alias{g.c,THETA-method}
\alias{g.d,THETA-method}
\alias{g.pdf,THETA-method}
\alias{g.Theta,THETA-method}
\alias{g.Theta1<-,THETA-method}
\alias{g.Theta2<-,THETA-method}
\title{
?
}
\description{
?
}
\keyword{random number generation}


