\name{truck}
\docType{data}
\alias{truck}
\title{Truck Dataset}
\description{
The dataset contains amplitudes and means measured on a truck wheels.
}
\usage{
data(truck)
}
\format{
\code{truck} is a data frame with 31665 rows and 2 variables (columns) named:
\enumerate{
\item{
\code{Amplitude} continuous.
}
\item{
\code{Mean} continuous.
}
}
}
\author{Mitja Franko}
\examples{
data(truck)
}
\keyword{datasets}
