\name{SSE}
\alias{SSE}
\title{
Sum of Squares Error
}
\description{
Returns the sum of squares error at \code{pos} for class REBMIX.
}
\usage{
SSE(x, pos = 1, ...)
}
\arguments{
\item{x}{
an object of class \code{REBMIX}.
}
\item{pos}{
a desired row number in \code{summary} for which the information criterion is calculated. The default value is \code{1}.
}
\item{\dots}{
currently not used.
}
}
\references{
C. M. Bishop. Neural Networks for Pattern Recognition. Clarendon Press, Oxford, 1995.
}
\keyword{information criterion}
