\name{weibull}
\docType{data}
\alias{weibull}
\title{Weibull Dataset 8.1}
\description{
The complete data are the failure times in weeks.
}
\usage{
weibull
}
\format{
\code{weibull} is a data frame with 50 cases (rows) and 1 variables (columns) named:
\enumerate{
\item{
\code{Failure.Time} continuous.
}
}
}
\references{
Murthy DNP, Jiang MXR (2004). \emph{Weibull Models}. John Wiley & Sons, New York.
}
\examples{
data("weibull")
}
\keyword{datasets}
