\name{weibullnormal}
\docType{data}
\alias{weibullnormal}
\title{Weibull-normal Simulated Dataset}
\description{
The dataset contains amplitudes and means simulated from a three component Weibull-normal mixture.
}
\usage{
data(weibullnormal)
}
\format{
\code{weibullnormal} is a data frame with 10000 rows and 2 variables (columns) named:
\enumerate{
\item{
\code{Amplitude} continuous.
}
\item{
\code{Mean} continuous.
}
}
}
\author{Mitja Franko}
\examples{
data(weibullnormal)
}
\keyword{datasets}
